全功能使能，修复bug，删除多余功能

+++ Inc/main.h
@@ -73,7 +73,7 @@ uint8_t GY30_CHRM;
 void Error_Handler(void);
 
 /* USER CODE BEGIN EFP */
-void delay(uint8_t k);
+void delay(uint32_t k);
 /* USER CODE END EFP */
 
 /* Private defines -----------------------------------------------------------*/
